Release/1.3.0

+++ app/Http/Controllers/OrderController.php
@@ -20,23 +20,27 @@ class OrderController extends Controller
     public function show(): View
     {
         $order = $this->orderService->getOrderByUser();
+        $this->authorize('update', $order);
         return view('store.orders.show', ['order' => $order]);
     }
 
     public function deleteOrder(): RedirectResponse
     {
+        $this->authorize('can', 'store_update_order');
         $this->orderService->deleteOrder();
         return redirect()->route(self::$defaultView);
     }
 
     public function addProduct(AddProductOrderRequest $request, Product $product): RedirectResponse
     {
+        $this->authorize('can', 'store_update_order');
         $this->orderService->addOrUpdateProduct($request, $product);
         return redirect()->route(self::$defaultView);
     }
 
     public function removeProduct(Product $product): RedirectResponse
     {
+        $this->authorize('can', 'store_update_order');
         $this->orderService->removeProduct($product);
         return redirect()->route(self::$defaultView);
     }
